services/miniflux-service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
databases==0.8.0
distlib==0.3.8
fastapi==0.110.0
filelock==3.13.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
miniflux==1.0.0
platformdirs==4.2.0
psycopg2==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==1.4.51
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1

#for testing       #new
pytest
requests